Add docset `products` attribute
