+++ docs/syntax/code.md
@@ -1,3 +1,8 @@
+---
+products:
+  - apm
+---
+
 # Code
 
 Code blocks can be used to display multiple lines of code. They preserve formatting and provide syntax highlighting when possible.
