+++ src/Elastic.Documentation.Configuration/Suggestions/Suggestions.cs
@@ -2,7 +2,7 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-namespace Elastic.Markdown.Suggestions;
+namespace Elastic.Documentation.Configuration.Suggestions;
 
 public class Suggestion(IReadOnlySet<string> candidates, string input)
 {
