+++ src/Elastic.Markdown/Slices/Index.cshtml
@@ -23,7 +23,7 @@
 		StaticFileContentHashProvider = Model.StaticFileContentHashProvider,
 		ReportIssueUrl = Model.ReportIssueUrl,
 		LegacyPage = Model.LegacyPage,
-		Products = Model.CurrentDocument.YamlFrontMatter?.Products is { Count: > 0} products ? string.Join(",", products.Select(p => p.DisplayName)) : null,
+		Products = Model.Products is { Count: > 0} products ? string.Join(",", products.Select(p => p.DisplayName)) : null,
 	};
 }
 <section id="elastic-docs-v3">
